add to readme, separate app sidebar/menubar components
